fix: remove circular deps
