+++ src/help/index.ts
@@ -4,7 +4,7 @@ import stripAnsi from 'strip-ansi'
 import {colorize} from '../cli-ux/theme'
 import write from '../cli-ux/write'
 import {Command} from '../command'
-import {error} from '../errors'
+import {error} from '../errors/error'
 import * as Interfaces from '../interfaces'
 import {load} from '../module-loader'
 import {SINGLE_COMMAND_CLI_SYMBOL} from '../symbols'
